requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.1
lxml==5.1.0
python-telegram-bot==13.7
